Fixing alignment issue in SDATA2 section
- This was resulting an unaligned access to SDATA2 aread from
    p10_sbe_check_quiesce(), during MPIPL
- This was causing caused by a const uint8_t array of size 3.
- changed array type to uint32_t to fix this

CQ: SW525033
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/116544
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>

+++ src/hwpf/target.C
@@ -1096,7 +1096,7 @@ fapi_try_exit:
             PMIC_TARGET_OFFSET,
             GI2C_TARGET_OFFSET
         };
-        const uint8_t PPE_TARGET_TYPE_ARRAY[] = {
+        const uint32_t PPE_TARGET_TYPE_ARRAY[] = {
             PPE_TARGET_TYPE_OCMB,
             PPE_TARGET_TYPE_PMIC,
             PPE_TARGET_TYPE_GENERICI2CSLAVE
